StaticMeshVertex, StaticMeshModel 클래스 추가 일부 멤버 REFLECTION 변경 기본 모델 변경 FBXLoader 제거

+++ XBOXController.h
@@ -52,8 +52,8 @@ namespace DK
 			ZeroMemory(&Vibration, sizeof(XINPUT_VIBRATION));
 
 			// Set the Vibration Values
-			Vibration.wLeftMotorSpeed = leftVal;
-			Vibration.wRightMotorSpeed = rightVal;
+			Vibration.wLeftMotorSpeed = static_cast<WORD>(leftVal);
+			Vibration.wRightMotorSpeed = static_cast<WORD>(rightVal);
 
 			// Vibrate the controller
 			XInputSetState(_playerNumber, &Vibration);
